feat(script): Add add function
Added an add function that adds numbers a and b to each other.
This function is used instead of eval, as eval is prone to injection
attacks. Add provides a safer alternative. This function is going to be
used to evaluate the addition of two numbers in the calculator program.